minor error handling fix
