+++ backend/workflow_manager/endpoint/destination.py
@@ -27,7 +27,6 @@ from workflow_manager.endpoint.exceptions import (
     InvalidToolOutputType,
     MissingDestinationConnectionType,
     ToolOutputTypeMismatch,
-    UnstractFSException,
 )
 from workflow_manager.endpoint.models import WorkflowEndpoint
 from workflow_manager.endpoint.queue_utils import QueueResult, QueueUtils
@@ -37,7 +36,8 @@ from workflow_manager.workflow.file_history_helper import FileHistoryHelper
 from workflow_manager.workflow.models.file_history import FileHistory
 from workflow_manager.workflow.models.workflow import Workflow
 
-from unstract.connectors.exceptions import ConnectorError
+from backend.exceptions import UnstractFSException
+from unstract.connectors.exceptions import AzureInvalidDirectoryError
 
 logger = logging.getLogger(__name__)
 
@@ -248,7 +248,7 @@ class DestinationConnector(BaseConnector):
                     destination_fs.upload_file_to_storage(
                         source_path=source_path, destination_path=destination_path
                     )
-        except ConnectorError as e:
+        except AzureInvalidDirectoryError as e:
             raise UnstractFSException(core_err=e) from e
 
     def insert_into_db(self, input_file_path: str) -> None:
